Create Arranging of flower sticks in a bouquet problem
Emma wants to gift a bouquet to her father on his birthday and asked for help from mother Rosy. Rosy gives n flower sticks numbered 1 to n to Emma and tells her to arrange it in the bouquet in a particular order. She asks her to arrange the first k flower sticks in the order of their increasing length and the remaining sticks in an order of their decreasing length.
Example:
Input: 34672981
Output :  12349876