Create Parking lot problem
A parking lot in a mall has R*C  number of parking  spaces. Each parking space will  either be empty(0) or full(1). The status (0/1) of  a parking space is represented as the element  of the matrix. The task is to find the row (R) in the parking lot that has the most of the space full(1)
Here, R*C is the size of the matrix. 
The elements  of the matrix should be only 1 or 0.

Example : 
Input : [[0, 1, 0],[ 1, 1, 0 ],[1, 0, 1],[ 1, 1, 1]]
Output: 4