Create Count maximum number of guests
A party has been organised on a cruise. The party is organised for a limited time T which is same as length of the arrays.The number of guests entering(E[i]) and number of guests leaving(L[i]) the party at every hour is represented as elements of the array. The task is to find the maximum number of guests  present on the cruise at any given instant  within T hours.
Input : E[] -> [7,0,5,1,3] & L[] -> [1,2,1,3,4]
Output: 8 (Maximum number of guests at an instance)
Explanation : 
1st hour : 
entry : 7 exit : 1
No. of guests on ship : 6
2nd hour : 
entry : 0 exit : 2
No. of guests on ship : 6-2=4
Hour 3:
entry : 5 exit : 1
No. of guests on ship : 4+5-1=8
Hour 4:
entry : 1 exit : 3
No. of guests on ship : 8+1-3=6
Hour 5:
entry : 3 exit : 4
No. of guests on ship : 6+3-4=5
Hence, number of guests within 5 hours is 8.