Create Balloon Problem
At a fun fair, a street vendor is selling different colours of balloons. He sells  N number of different colours of balloons(B[]). The task is to find the colour(odd) of the balloon which is present odd number of times in the bunch of balloons.
Note:
If there is more than one colour which is odd in number , then the colour in the array which is present odd number of times is displayed. The colours of the balloons can all be either upper case or lower case in the list. 
Example:
Input: ['r','g','b','b','g','y','y','g']
Output: r

Example:
Input : list1=['r','g','b','b','g','y','y','g','r','g']
Output : All are even